+++ apps/mobile-dev-client/app/core-webview.tsx
@@ -1,3 +1,0 @@
-import { CoreWebview } from "@seoulcomix/core-mobile";
-
-export default () => <CoreWebview isScreen />;