+++ example/sample-expo/App.tsx
@@ -1,6 +1,6 @@
 import { StyleSheet, Text, View } from "react-native";
 import { StatusBar } from "expo-status-bar";
-import { Button } from "@seoulcomix/ui";
+import { Button } from "@expo-bridge/ui";
 
 export default function Native() {
   return (