+++ packages/core-bridge/src/index.ts
@@ -213,3 +213,5 @@ export const useCoreBridgeListener = (
     };
   }, [cb]);
 };
+
+export * from "./provider";