+++ packages/core-mobile/src/providers/index.ts
@@ -1,2 +1,2 @@
-export { CoreProviders } from "./core";
+export { CoreMobileProvider } from "./core";
 export { useCoreContext } from "./option-context";