+++ packages/core-storage/CHANGELOG.md
@@ -1,14 +1,14 @@
-# @seoulcomix/core-navigation
+# @expo-bridge/core-navigation
 
 ## 0.1.0
 
 ### Patch Changes
 
-- @seoulcomix/core-bridge@0.1.0
+- @expo-bridge/core-bridge@0.1.0
 
 ## 0.0.1
 
 ### Patch Changes
 
 - 2cd4942: feat: 🎸 add readme
-  - @seoulcomix/core-bridge@0.0.1
+  - @expo-bridge/core-bridge@0.0.1